Add files via upload
Develop a system using UML for E-Book Management System. It should be designed to manage the 
books that were read through the internet. That should consist of the details of the e-book that were 
read by the user online. It will be controlled by the central system. This system should act as a 
backup of all details together